requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
python-dotenv
psycopg2
